Fix import issues by porting vector back as subclass of matrix.py
